Store the hard-coded '1' for replication as a short by casting it

+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/serverless/invoking/ServerlessNameNodeClient.java
@@ -1283,7 +1283,7 @@ public class ServerlessNameNodeClient implements ClientProtocol {
         opArguments.put("enumSetBase64", enumSetBase64);
         opArguments.put("createParent", createParent);
         LOG.warn("Using hard-coded replication value of 1.");
-        opArguments.put("replication", 1);
+        opArguments.put("replication", (short)1);
         opArguments.put("blockSize", blockSize);
 
         // Include a flag to indicate whether the policy is non-null.
